requirements.txt
Django==3.2
djangorestframework==3.12.4
gunicorn==20.1.0
python-decouple==3.4
